Step1: Move pilot shared headers into `common/`
The pilot C++ core is a flat directory of about forty translation
units, and the only structure is the file name.  This starts the
reorganization into per-neighborhood subdirectories by moving the
shared base headers, which nearly every pilot source includes, into
`common/`: `common_detail.hpp`, `platform.hpp`, and `render_misc`.

Moving the most-included headers first exercises the include-path
mechanics against nearly every file before the larger neighborhoods
move.  All includes are angle-bracket and rooted at the package, so
every include site across the pilot updates in lockstep, and the
CMake source lists change paths only.  No logic changes.
