+++ cpp/solvcon/pilot/R2DWidget.hpp
@@ -12,7 +12,7 @@
  * @ingroup group_domain
  */
 
-#include <solvcon/pilot/common_detail.hpp> // Must be the first include.
+#include <solvcon/pilot/common/common_detail.hpp> // Must be the first include.
 
 #include <solvcon/buffer/small_vector.hpp>
 #include <solvcon/pilot/DrawTool.hpp>
