+++ cpp/solvcon/pilot/RAction.hpp
@@ -12,7 +12,7 @@
  * @ingroup group_domain
  */
 
-#include <solvcon/pilot/common_detail.hpp> // Must be the first include.
+#include <solvcon/pilot/common/common_detail.hpp> // Must be the first include.
 
 #include <QAction>
