+++ cpp/solvcon/pilot/RDomainWidget.hpp
@@ -13,7 +13,7 @@
  * @ingroup group_domain
  */
 
-#include <solvcon/pilot/common_detail.hpp> // Must be the first include.
+#include <solvcon/pilot/common/common_detail.hpp> // Must be the first include.
 
 #include <solvcon/pilot/RAxisGizmo.hpp>
 #include <solvcon/pilot/RColormap.hpp>
