+++ cpp/solvcon/pilot/RMaterial.hpp
@@ -13,7 +13,7 @@
  * @ingroup group_domain
  */
 
-#include <solvcon/pilot/common_detail.hpp> // Must be the first include.
+#include <solvcon/pilot/common/common_detail.hpp> // Must be the first include.
 
 #include <rhi/qrhi.h>
