+++ cpp/solvcon/pilot/RShortcutManager.hpp
@@ -17,7 +17,7 @@
  * @ingroup group_domain
  */
 
-#include <solvcon/pilot/common_detail.hpp> // Must be the first include.
+#include <solvcon/pilot/common/common_detail.hpp> // Must be the first include.
 
 #include <string>
 #include <vector>
