+++ cpp/solvcon/pilot/common/common_detail.hpp
