+++ cpp/solvcon/pilot/common/platform.hpp
