+++ cpp/solvcon/pilot/theme.hpp
@@ -23,7 +23,7 @@
 
 #include <cstdint>
 
-#include <solvcon/pilot/platform.hpp>
+#include <solvcon/pilot/common/platform.hpp>
 
 namespace solvcon
 {
